move test to right class
